requirements.txt
requests>=2.8.1
urllib3>=2.2.2
pydantic[email]>=2.8.2
fastmcp>=2.13.0.2
eunomia-mcp>=0.3.10
pydantic-ai-slim[fastmcp,openai,anthropic,google,huggingface,a2a,ag-ui,web]>=1.32.0
